Pizza Size implemented.

+++ ProductManagementSystem/src/main/java/com/ferozkhandev/PizzaOrderingSystem/ProductManagementSystem/repositories/PizzaSizeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PizzaSizeRepository extends JpaRepository<PizzaSize, Integer> {
+public interface PizzaSizeRepository extends JpaRepository<PizzaSize, Long> {
 }
